Simplicy some logic + separate resizing logic in its own hook
