+++ packages/toolpad-app/src/toolpad/AppEditor/PageEditor/RenderPanel/RenderPanel.tsx
@@ -5,9 +5,9 @@ import DeleteIcon from '@mui/icons-material/Delete';
 import { IconButton, styled } from '@mui/material';
 import { RuntimeEvent, NodeId } from '@mui/toolpad-core';
 import { useNavigate } from 'react-router-dom';
-import { FlowDirection, NodeInfo, SlotsState, SlotState } from '../../../types';
-import * as appDom from '../../../appDom';
-import EditorCanvasHost, { EditorCanvasHostHandle } from './EditorCanvasHost';
+import { FlowDirection, NodeInfo, SlotsState, SlotState } from '../../../../types';
+import * as appDom from '../../../../appDom';
+import EditorCanvasHost, { EditorCanvasHostHandle } from '../EditorCanvasHost';
 import {
   absolutePositionCss,
   getRectanglePointEdge,
@@ -18,10 +18,10 @@ import {
   RECTANGLE_EDGE_RIGHT,
   RECTANGLE_EDGE_TOP,
   rectContainsPoint,
-} from '../../../utils/geometry';
-import { PinholeOverlay } from '../../../PinholeOverlay';
-import { getPageViewState } from '../../../pageViewState';
-import { useDom, useDomApi } from '../../DomLoader';
+} from '../../../../utils/geometry';
+import { PinholeOverlay } from '../../../../PinholeOverlay';
+import { getPageViewState } from '../../../../pageViewState';
+import { useDom, useDomApi } from '../../../DomLoader';
 import {
   DropZone,
   DROP_ZONE_BOTTOM,
@@ -31,8 +31,8 @@ import {
   DROP_ZONE_TOP,
   usePageEditorApi,
   usePageEditorState,
-} from './PageEditorProvider';
-import { useToolpadComponent } from '../toolpadComponents';
+} from '../PageEditorProvider';
+import { useToolpadComponent } from '../../toolpadComponents';
 import {
   getElementNodeComponentId,
   PAGE_COLUMN_COMPONENT_ID,
@@ -40,9 +40,10 @@ import {
   isPageLayoutComponent,
   isPageRow,
   isPageColumn,
-} from '../../../toolpadComponents';
-import { ExactEntriesOf } from '../../../utils/types';
-import { OverlayGrid } from './OverlayGrid';
+} from '../../../../toolpadComponents';
+import { ExactEntriesOf } from '../../../../utils/types';
+import { OverlayGrid, OverlayGridHandle } from './OverlayGrid';
+import { useNodeResize } from './useNodeResize';
 
 const classes = {
   view: 'Toolpad_View',
@@ -534,12 +535,7 @@ export default function RenderPanel({ className }: RenderPanelProps) {
     return [pageNode, ...appDom.getDescendants(dom, pageNode)];
   }, [dom, pageNode]);
 
-  const overlayGridRef = React.useRef<{
-    gridElement: HTMLDivElement | null;
-    getMinColumnWidth: () => number;
-    getLeftColumnEdges: () => number[];
-    getRightColumnEdges: () => number[];
-  }>({
+  const overlayGridRef = React.useRef<OverlayGridHandle>({
     gridElement: null,
     getMinColumnWidth: () => 0,
     getLeftColumnEdges: () => [],
@@ -825,84 +821,6 @@ export default function RenderPanel({ className }: RenderPanelProps) {
 
     return rects;
   }, [nodesInfo, pageNodes]);
-
-  const handleEdgeDragStart = React.useCallback(
-    (node: appDom.ElementNode, edge: RectangleEdge) => (event: React.MouseEvent<HTMLElement>) => {
-      event.stopPropagation();
-
-      api.edgeDragStart({ nodeId: node.id, edge });
-
-      api.select(node.id);
-    },
-    [api],
-  );
-
-  const resizePreviewElementRef = React.useRef<HTMLDivElement | null>(null);
-  const resizePreviewElement = resizePreviewElementRef.current;
-
-  const handleEdgeDragOver = React.useCallback(
-    (event: React.MouseEvent<Element>) => {
-      const draggedNode = getCurrentlyDraggedNode();
-
-      if (!draggedNode) {
-        return;
-      }
-
-      const draggedNodeInfo = nodesInfo[draggedNode.id];
-      const draggedNodeRect = draggedNodeInfo?.rect;
-
-      const parent = draggedNode && appDom.getParent(dom, draggedNode);
-
-      const parentInfo = parent ? nodesInfo[parent.id] : null;
-      const parentRect = parentInfo?.rect;
-
-      const cursorPos = canvasHostRef.current?.getViewCoordinates(event.clientX, event.clientY);
-
-      if (draggedNodeRect && parentRect && resizePreviewElement && cursorPos) {
-        const resizeSnapUnits = 4; // px
-        const snapToGridMargin = 10; // px
-
-        let snappedToGridCursorPosX = Math.round(cursorPos.x / resizeSnapUnits) * resizeSnapUnits;
-
-        const activeSnapGridColumnEdges =
-          draggedEdge === RECTANGLE_EDGE_LEFT
-            ? overlayGridRef.current.getLeftColumnEdges()
-            : overlayGridRef.current.getRightColumnEdges();
-
-        for (const gridColumnEdge of activeSnapGridColumnEdges) {
-          if (Math.abs(gridColumnEdge - cursorPos.x) <= snapToGridMargin) {
-            snappedToGridCursorPosX = gridColumnEdge;
-          }
-        }
-
-        const minGridColumnWidth = overlayGridRef.current.getMinColumnWidth();
-
-        if (
-          draggedEdge === RECTANGLE_EDGE_LEFT &&
-          cursorPos.x > parentRect.x + minGridColumnWidth &&
-          cursorPos.x < draggedNodeRect.x + draggedNodeRect.width - minGridColumnWidth
-        ) {
-          const updatedTransformScale =
-            1 + (draggedNodeRect.x - snappedToGridCursorPosX) / draggedNodeRect.width;
-
-          resizePreviewElement.style.transformOrigin = '100% 50%';
-          resizePreviewElement.style.transform = `scale(${updatedTransformScale}, 1)`;
-        }
-        if (
-          draggedEdge === RECTANGLE_EDGE_RIGHT &&
-          cursorPos.x > draggedNodeRect.x + minGridColumnWidth &&
-          cursorPos.x < parentRect.x + parentRect.width - minGridColumnWidth
-        ) {
-          const updatedTransformScale =
-            (snappedToGridCursorPosX - draggedNodeRect.x) / draggedNodeRect.width;
-
-          resizePreviewElement.style.transformOrigin = '0 50%';
-          resizePreviewElement.style.transform = `scale(${updatedTransformScale}, 1)`;
-        }
-      }
-    },
-    [dom, draggedEdge, getCurrentlyDraggedNode, nodesInfo, resizePreviewElement],
-  );
 
   const handleNodeDragOver = React.useCallback(
     (event: React.DragEvent<Element>) => {
@@ -1527,149 +1445,6 @@ export default function RenderPanel({ className }: RenderPanelProps) {
     ],
   );
 
-  const normalizePageRowColumnSizes = React.useCallback(
-    (pageRowNode: appDom.ElementNode): number[] => {
-      const children = appDom.getChildNodes(dom, pageRowNode).children;
-
-      const layoutColumnSizes = children.map((child) => child.layout?.columnSize?.value || 1);
-      const totalLayoutColumnSizes = layoutColumnSizes.reduce((acc, size) => acc + size, 0);
-
-      const normalizedLayoutColumnSizes = layoutColumnSizes.map(
-        (size) => (size * children.length) / totalLayoutColumnSizes,
-      );
-
-      children.forEach((child, childIndex) => {
-        if (child.layout?.columnSize) {
-          domApi.setNodeNamespacedProp(
-            child,
-            'layout',
-            'columnSize',
-            appDom.createConst(normalizedLayoutColumnSizes[childIndex]),
-          );
-        }
-      });
-
-      return normalizedLayoutColumnSizes;
-    },
-    [dom, domApi],
-  );
-
-  const previousRowColumnCountsRef = React.useRef<Record<NodeId, number>>({});
-
-  React.useEffect(() => {
-    pageNodes.forEach((node: appDom.AppDomNode) => {
-      if (appDom.isElement(node) && isPageRow(node)) {
-        const children = appDom.getChildNodes(dom, node).children;
-        const childrenCount = children.length;
-
-        if (childrenCount < previousRowColumnCountsRef.current[node.id]) {
-          normalizePageRowColumnSizes(node);
-        }
-
-        previousRowColumnCountsRef.current[node.id] = childrenCount;
-      }
-    });
-  }, [dom, normalizePageRowColumnSizes, pageNodes]);
-
-  const handleEdgeDragEnd = React.useCallback(
-    (event: React.MouseEvent<Element>) => {
-      event.preventDefault();
-
-      const draggedNode = getCurrentlyDraggedNode();
-
-      if (!draggedNode) {
-        return;
-      }
-
-      const draggedNodeInfo = nodesInfo[draggedNode.id];
-      const draggedNodeRect = draggedNodeInfo?.rect;
-
-      const parent = appDom.getParent(dom, draggedNode);
-
-      const parentChildren = parent ? appDom.getChildNodes(dom, parent).children : [];
-      const totalLayoutColumnSizes = parentChildren.reduce(
-        (acc, child) => acc + (nodesInfo[child.id]?.rect?.width || 0),
-        0,
-      );
-
-      const resizePreviewRect = resizePreviewElement?.getBoundingClientRect();
-
-      if (draggedNodeRect && resizePreviewRect) {
-        if (draggedEdge === RECTANGLE_EDGE_LEFT) {
-          const previousSibling = appDom.getSiblingBeforeNode(dom, draggedNode, 'children');
-
-          if (previousSibling) {
-            const previousSiblingInfo = nodesInfo[previousSibling.id];
-            const previousSiblingRect = previousSiblingInfo?.rect;
-
-            if (previousSiblingRect) {
-              const updatedDraggedNodeColumnSize = Math.max(
-                0,
-                (resizePreviewRect.width * parentChildren.length) / totalLayoutColumnSizes,
-              );
-              const updatedPreviousSiblingColumnSize = Math.max(
-                0,
-                ((previousSiblingRect.width - (resizePreviewRect.width - draggedNodeRect.width)) *
-                  parentChildren.length) /
-                  totalLayoutColumnSizes,
-              );
-
-              domApi.setNodeNamespacedProp(
-                draggedNode,
-                'layout',
-                'columnSize',
-                appDom.createConst(updatedDraggedNodeColumnSize),
-              );
-              domApi.setNodeNamespacedProp(
-                previousSibling,
-                'layout',
-                'columnSize',
-                appDom.createConst(updatedPreviousSiblingColumnSize),
-              );
-            }
-          }
-        }
-        if (draggedEdge === RECTANGLE_EDGE_RIGHT) {
-          const nextSibling = appDom.getSiblingAfterNode(dom, draggedNode, 'children');
-
-          if (nextSibling) {
-            const nextSiblingInfo = nodesInfo[nextSibling.id];
-            const nextSiblingRect = nextSiblingInfo?.rect;
-
-            if (nextSiblingRect) {
-              const updatedDraggedNodeColumnSize = Math.max(
-                0,
-                (resizePreviewRect.width * parentChildren.length) / totalLayoutColumnSizes,
-              );
-              const updatedNextSiblingColumnSize = Math.max(
-                0,
-                ((nextSiblingRect.width - (resizePreviewRect.width - draggedNodeRect.width)) *
-                  parentChildren.length) /
-                  totalLayoutColumnSizes,
-              );
-
-              domApi.setNodeNamespacedProp(
-                draggedNode,
-                'layout',
-                'columnSize',
-                appDom.createConst(updatedDraggedNodeColumnSize),
-              );
-              domApi.setNodeNamespacedProp(
-                nextSibling,
-                'layout',
-                'columnSize',
-                appDom.createConst(updatedNextSiblingColumnSize),
-              );
-            }
-          }
-        }
-      }
-
-      api.dragEnd();
-    },
-    [api, dom, domApi, draggedEdge, getCurrentlyDraggedNode, nodesInfo, resizePreviewElement],
-  );
-
   const handleNodeDragEnd = React.useCallback(
     (event: DragEvent | React.DragEvent) => {
       event.preventDefault();
@@ -1805,6 +1580,14 @@ export default function RenderPanel({ className }: RenderPanelProps) {
     },
     [dom, domApi, api, navigate],
   );
+
+  const { handleEdgeDragStart, handleEdgeDragOver, handleEdgeDragEnd, resizePreviewElementRef } =
+    useNodeResize({
+      pageNodes,
+      getCurrentlyDraggedNode,
+      canvasHostRef,
+      overlayGridRef,
+    });
 
   return (
     <RenderPanelRoot className={className}>
