+++ packages/toolpad-app/src/toolpad/AppEditor/PageEditor/RenderPanel/index.ts
@@ -1,0 +1,3 @@
+import RenderPanel from './RenderPanel';
+
+export default RenderPanel;
